[FE] MOVE: 폴더명 변경을 위한 커밋  #39
