--- src/frontend/src/pages/friendPage/index.tsx
@@ -1,3 +1,0 @@
-export const FriendPage = () => {
-  return <div>Friend</div>;
-};
