--- src/frontend/src/pages/homePage/index.tsx
@@ -1,3 +1,0 @@
-export const HomePage = () => {
-  return <div>Home</div>;
-};
