--- src/frontend/src/pages/layout/index.css.ts
@@ -1,24 +1,0 @@
-import { styled } from 'styled-components';
-
-export const Wrapper = styled.div`
-  display: flex;
-  align-items: flex-start;
-  width: 100%;
-  height: 100%;
-`;
-
-export const Container = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-`;
-
-export const LinkContainer = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  gap: 20px;
-  padding: 20px;
-`;
