--- src/frontend/src/pages/loginPage/index.tsx
@@ -1,3 +1,0 @@
-export const LoginPage = () => {
-  return <div>Login</div>;
-};
