--- src/frontend/src/pages/myRoomPage/index.tsx
@@ -1,3 +1,0 @@
-export const MyRoomPage = () => {
-  return <div>내 방</div>;
-};
