--- src/frontend/src/pages/notFoundPage/index.tsx
@@ -1,3 +1,0 @@
-export const NotFoundPage = () => {
-  return <div>404 Not Found</div>;
-};
