--- src/frontend/src/pages/registerPage/index.tsx
@@ -1,3 +1,0 @@
-export const RegisterPage = () => {
-  return <div>회원가입</div>;
-};
