--- src/frontend/src/pages/settingPage/index.tsx
@@ -1,3 +1,0 @@
-export const SettingPage = () => {
-  return <div>설정</div>;
-};
